move badge filter to shared, apply to mobile
