+++ apps/web/src/constants/community.ts
@@ -1,16 +1,2 @@
 export const GRID_ITEM_PER_PAGE = 40;
 export const LIST_ITEM_PER_PAGE = 200;
-
-export const GRID_ENABLED_COMMUNITY_ADDRESSES = [
-  // Art Gobblers
-  '0x60bb1e2aa1c9acafb4d34f71585d7e959f387769',
-  // Monarch
-  '0xc729ce9bf1030fbb639849a96fa8bbd013680b64',
-  // Cryptocoven
-  '0x5180db8f5c931aae63c74266b211f580155ecac8',
-];
-
-export const BADGE_ENABLED_COMMUNITY_ADDRESSES = new Set([
-  // Gallery Premium Membership Cards
-  '0xe01569ca9b39e55bc7c0dfa09f05fa15cb4c7698',
-]);
